Refactor: E2E tests
