+++ operator/test/e2e/e2e_test.go
@@ -20,11 +20,8 @@ limitations under the License.
 package e2e
 
 import (
-	"encoding/json"
 	"fmt"
-	"os"
 	"os/exec"
-	"path/filepath"
 	"strings"
 	"time"
 
@@ -34,27 +31,26 @@ import (
 	"github.com/MirrorStudios/fallernetes-operator/operator/test/utils"
 )
 
-// namespace where the project is deployed in
+// namespace where the operator is deployed.
 const namespace = "operator-system"
 
-// serviceAccountName created for the project
+// serviceAccountName is the controller-manager service account.
 const serviceAccountName = "operator-controller-manager"
 
-// metricsServiceName is the name of the metrics service of the project
+// metricsServiceName is the metrics service of the operator.
 const metricsServiceName = "operator-controller-manager-metrics-service"
 
-// metricsRoleBindingName is the name of the RBAC that will be created to allow get the metrics data
+// metricsRoleBindingName is the RBAC binding created to scrape metrics.
 const metricsRoleBindingName = "operator-metrics-binding"
 
-// testNamespace is the namespace used for Fleet test workloads (distinct from the operator namespace).
+// testNamespace is the namespace used for game-server workloads.
 const testNamespace = "default"
 
 var _ = Describe("Manager", Ordered, func() {
+	// controllerPodName is resolved in the first infrastructure test and reused for
+	// log collection in AfterEach.
 	var controllerPodName string
 
-	// Before running the tests, set up the environment by creating the namespace,
-	// enforce the restricted security policy to the namespace, installing CRDs,
-	// and deploying the controller.
 	BeforeAll(func() {
 		By("creating manager namespace")
 		cmd := exec.Command("kubectl", "create", "ns", namespace)
@@ -78,19 +74,14 @@ var _ = Describe("Manager", Ordered, func() {
 		Expect(err).NotTo(HaveOccurred(), "Failed to deploy the controller-manager")
 	})
 
-	// After all tests have been executed, clean up by undeploying the controller, uninstalling CRDs,
-	// and deleting the namespace.
 	AfterAll(func() {
-		// In Ginkgo v2, DeferCleanup registered in the last It of an Ordered container runs after
-		// AfterAll, not before it. Explicitly delete all remaining fleets here so AfterAll does not
-		// depend on DeferCleanup having already submitted the delete requests.
+		// Delete all fleets before undeploying — finalizers prevent namespace deletion
+		// when the controller is gone.
 		By("deleting any remaining fleets in default namespace")
 		cmd := exec.Command("kubectl", "delete", "fleet", "--all", "-n", "default",
 			"--wait=false", "--ignore-not-found=true")
 		_, _ = utils.Run(cmd)
 
-		// Wait for fleet resources to be fully deleted (finalizers removed) before undeploying.
-		// Undeploying while finalizers are pending blocks kubectl delete (waiting on webhook/namespace).
 		By("waiting for all fleets in default namespace to be fully deleted")
 		Eventually(func(g Gomega) {
 			cmd := exec.Command("kubectl", "get", "fleets",
@@ -102,7 +93,8 @@ var _ = Describe("Manager", Ordered, func() {
 		}, 3*time.Minute, 5*time.Second).Should(Succeed())
 
 		By("cleaning up the curl pod for metrics")
-		cmd = exec.Command("kubectl", "delete", "pod", "curl-metrics", "-n", namespace)
+		cmd = exec.Command("kubectl", "delete", "pod", "curl-metrics", "-n", namespace,
+			"--ignore-not-found=true")
 		_, _ = utils.Run(cmd)
 
 		By("undeploying the controller-manager")
@@ -118,66 +110,53 @@ var _ = Describe("Manager", Ordered, func() {
 		_, _ = utils.Run(cmd)
 	})
 
-	// After each test, check for failures and collect logs, events,
-	// and pod descriptions for debugging.
 	AfterEach(func() {
-		specReport := CurrentSpecReport()
-		if specReport.Failed() {
-			By("Fetching controller manager pod logs")
-			cmd := exec.Command("kubectl", "logs", controllerPodName, "-n", namespace)
-			controllerLogs, err := utils.Run(cmd)
-			if err == nil {
-				_, _ = fmt.Fprintf(GinkgoWriter, "Controller logs:\n %s", controllerLogs)
-			} else {
-				_, _ = fmt.Fprintf(GinkgoWriter, "Failed to get Controller logs: %s", err)
-			}
+		if !CurrentSpecReport().Failed() {
+			return
+		}
+		By("Fetching controller manager pod logs")
+		cmd := exec.Command("kubectl", "logs", controllerPodName, "-n", namespace)
+		if logs, err := utils.Run(cmd); err == nil {
+			_, _ = fmt.Fprintf(GinkgoWriter, "Controller logs:\n%s", logs)
+		} else {
+			_, _ = fmt.Fprintf(GinkgoWriter, "Failed to get controller logs: %s\n", err)
+		}
 
-			By("Fetching Kubernetes events")
-			cmd = exec.Command("kubectl", "get", "events", "-n", namespace, "--sort-by=.lastTimestamp")
-			eventsOutput, err := utils.Run(cmd)
-			if err == nil {
-				_, _ = fmt.Fprintf(GinkgoWriter, "Kubernetes events:\n%s", eventsOutput)
-			} else {
-				_, _ = fmt.Fprintf(GinkgoWriter, "Failed to get Kubernetes events: %s", err)
-			}
+		By("Fetching Kubernetes events")
+		cmd = exec.Command("kubectl", "get", "events", "-n", namespace, "--sort-by=.lastTimestamp")
+		if events, err := utils.Run(cmd); err == nil {
+			_, _ = fmt.Fprintf(GinkgoWriter, "Kubernetes events:\n%s", events)
+		} else {
+			_, _ = fmt.Fprintf(GinkgoWriter, "Failed to get Kubernetes events: %s\n", err)
+		}
 
-			By("Fetching curl-metrics logs")
-			cmd = exec.Command("kubectl", "logs", "curl-metrics", "-n", namespace)
-			metricsOutput, err := utils.Run(cmd)
-			if err == nil {
-				_, _ = fmt.Fprintf(GinkgoWriter, "Metrics logs:\n %s", metricsOutput)
-			} else {
-				_, _ = fmt.Fprintf(GinkgoWriter, "Failed to get curl-metrics logs: %s", err)
-			}
+		By("Fetching curl-metrics logs")
+		cmd = exec.Command("kubectl", "logs", "curl-metrics", "-n", namespace)
+		if out, err := utils.Run(cmd); err == nil {
+			_, _ = fmt.Fprintf(GinkgoWriter, "Metrics logs:\n%s", out)
+		}
 
-			By("Fetching controller manager pod description")
-			cmd = exec.Command("kubectl", "describe", "pod", controllerPodName, "-n", namespace)
-			podDescription, err := utils.Run(cmd)
-			if err == nil {
-				fmt.Println("Pod description:\n", podDescription)
-			} else {
-				fmt.Println("Failed to describe controller pod")
-			}
+		By("Fetching controller manager pod description")
+		cmd = exec.Command("kubectl", "describe", "pod", controllerPodName, "-n", namespace)
+		if desc, err := utils.Run(cmd); err == nil {
+			_, _ = fmt.Fprintf(GinkgoWriter, "Pod description:\n%s", desc)
 		}
 	})
 
 	SetDefaultEventuallyTimeout(2 * time.Minute)
 	SetDefaultEventuallyPollingInterval(time.Second)
 
-	Context("Manager", func() {
+	Context("infrastructure", func() {
 		It("should run successfully", func() {
 			By("validating that the controller-manager pod is running as expected")
-			verifyControllerUp := func(g Gomega) {
-				By("getting the name of the controller-manager pod")
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get",
 					"pods", "-l", "control-plane=controller-manager",
 					"-o", "go-template={{ range .items }}"+
 						"{{ if not .metadata.deletionTimestamp }}"+
-						"{{ .metadata.name }}"+
-						"{{ \"\\n\" }}{{ end }}{{ end }}",
+						"{{ .metadata.name }}{{ \"\\n\" }}{{ end }}{{ end }}",
 					"-n", namespace,
 				)
-
 				podOutput, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred(), "Failed to retrieve controller-manager pod information")
 				podNames := utils.GetNonEmptyLines(podOutput)
@@ -185,7 +164,6 @@ var _ = Describe("Manager", Ordered, func() {
 				controllerPodName = podNames[0]
 				g.Expect(controllerPodName).To(ContainSubstring("controller-manager"))
 
-				By("validating the pod's status")
 				cmd = exec.Command("kubectl", "get",
 					"pods", controllerPodName, "-o", "jsonpath={.status.phase}",
 					"-n", namespace,
@@ -193,8 +171,7 @@ var _ = Describe("Manager", Ordered, func() {
 				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred())
 				g.Expect(output).To(Equal("Running"), "Incorrect controller-manager pod status")
-			}
-			Eventually(verifyControllerUp).Should(Succeed())
+			}).Should(Succeed())
 		})
 
 		It("should ensure the metrics endpoint is serving metrics", func() {
@@ -217,57 +194,52 @@ var _ = Describe("Manager", Ordered, func() {
 			Expect(token).NotTo(BeEmpty())
 
 			By("ensuring the controller pod is ready")
-			verifyControllerPodReady := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get", "pod", controllerPodName, "-n", namespace,
 					"-o", "jsonpath={.status.conditions[?(@.type=='Ready')].status}")
 				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred())
 				g.Expect(output).To(Equal("True"), "Controller pod not ready")
-			}
-			Eventually(verifyControllerPodReady, 3*time.Minute, time.Second).Should(Succeed())
+			}, 3*time.Minute, time.Second).Should(Succeed())
 
 			By("verifying that the controller manager is serving the metrics server")
-			verifyMetricsServerStarted := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "logs", controllerPodName, "-n", namespace)
 				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred())
 				g.Expect(output).To(ContainSubstring("Serving metrics server"),
 					"Metrics server not yet started")
-			}
-			Eventually(verifyMetricsServerStarted, 3*time.Minute, time.Second).Should(Succeed())
+			}, 3*time.Minute, time.Second).Should(Succeed())
 
 			By("waiting for the webhook service endpoints to be ready")
-			verifyWebhookEndpointsReady := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get", "endpointslices.discovery.k8s.io", "-n", namespace,
 					"-l", "kubernetes.io/service-name=operator-webhook-service",
 					"-o", "jsonpath={range .items[*]}{range .endpoints[*]}{.addresses[*]}{end}{end}")
 				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred(), "Webhook endpoints should exist")
 				g.Expect(output).ShouldNot(BeEmpty(), "Webhook endpoints not yet ready")
-			}
-			Eventually(verifyWebhookEndpointsReady, 3*time.Minute, time.Second).Should(Succeed())
+			}, 3*time.Minute, time.Second).Should(Succeed())
 
 			By("verifying the mutating webhook server is ready")
-			verifyMutatingWebhookReady := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get", "mutatingwebhookconfigurations.admissionregistration.k8s.io",
 					"operator-mutating-webhook-configuration",
 					"-o", "jsonpath={.webhooks[0].clientConfig.caBundle}")
 				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred(), "MutatingWebhookConfiguration should exist")
 				g.Expect(output).ShouldNot(BeEmpty(), "Mutating webhook CA bundle not yet injected")
-			}
-			Eventually(verifyMutatingWebhookReady, 3*time.Minute, time.Second).Should(Succeed())
+			}, 3*time.Minute, time.Second).Should(Succeed())
 
 			By("verifying the validating webhook server is ready")
-			verifyValidatingWebhookReady := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get", "validatingwebhookconfigurations.admissionregistration.k8s.io",
 					"operator-validating-webhook-configuration",
 					"-o", "jsonpath={.webhooks[0].clientConfig.caBundle}")
 				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred(), "ValidatingWebhookConfiguration should exist")
 				g.Expect(output).ShouldNot(BeEmpty(), "Validating webhook CA bundle not yet injected")
-			}
-			Eventually(verifyValidatingWebhookReady, 3*time.Minute, time.Second).Should(Succeed())
+			}, 3*time.Minute, time.Second).Should(Succeed())
 
 			By("waiting additional time for webhook server to stabilize")
 			time.Sleep(5 * time.Second)
@@ -291,14 +263,10 @@ var _ = Describe("Manager", Ordered, func() {
 							"securityContext": {
 								"readOnlyRootFilesystem": true,
 								"allowPrivilegeEscalation": false,
-								"capabilities": {
-									"drop": ["ALL"]
-								},
+								"capabilities": {"drop": ["ALL"]},
 								"runAsNonRoot": true,
 								"runAsUser": 1000,
-								"seccompProfile": {
-									"type": "RuntimeDefault"
-								}
+								"seccompProfile": {"type": "RuntimeDefault"}
 							}
 						}],
 						"serviceAccountName": "%s"
@@ -307,674 +275,64 @@ var _ = Describe("Manager", Ordered, func() {
 			_, err = utils.Run(cmd)
 			Expect(err).NotTo(HaveOccurred(), "Failed to create curl-metrics pod")
 
-			By("waiting for the curl-metrics pod to complete.")
-			verifyCurlUp := func(g Gomega) {
+			By("waiting for the curl-metrics pod to complete")
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get", "pods", "curl-metrics",
 					"-o", "jsonpath={.status.phase}",
 					"-n", namespace)
 				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred())
 				g.Expect(output).To(Equal("Succeeded"), "curl pod in wrong status")
-			}
-			Eventually(verifyCurlUp, 5*time.Minute).Should(Succeed())
+			}, 5*time.Minute).Should(Succeed())
 
 			By("getting the metrics by checking curl-metrics logs")
-			verifyMetricsAvailable := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				metricsOutput, err := getMetricsOutput()
 				g.Expect(err).NotTo(HaveOccurred(), "Failed to retrieve logs from curl pod")
 				g.Expect(metricsOutput).NotTo(BeEmpty())
 				g.Expect(metricsOutput).To(ContainSubstring("< HTTP/1.1 200 OK"))
-			}
-			Eventually(verifyMetricsAvailable, 2*time.Minute).Should(Succeed())
+			}, 2*time.Minute).Should(Succeed())
 		})
 
 		It("should provisioned cert-manager", func() {
 			By("validating that cert-manager has the certificate Secret")
-			verifyCertManager := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get", "secrets", "webhook-server-cert", "-n", namespace)
 				_, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred())
-			}
-			Eventually(verifyCertManager).Should(Succeed())
+			}).Should(Succeed())
 		})
 
 		It("should have CA injection for mutating webhooks", func() {
 			By("checking CA injection for mutating webhooks")
-			verifyCAInjection := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get",
 					"mutatingwebhookconfigurations.admissionregistration.k8s.io",
 					"operator-mutating-webhook-configuration",
 					"-o", "go-template={{ range .webhooks }}{{ .clientConfig.caBundle }}{{ end }}")
-				mwhOutput, err := utils.Run(cmd)
+				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(len(mwhOutput)).To(BeNumerically(">", 10))
-			}
-			Eventually(verifyCAInjection).Should(Succeed())
+				g.Expect(len(output)).To(BeNumerically(">", 10))
+			}).Should(Succeed())
 		})
 
 		It("should have CA injection for validating webhooks", func() {
 			By("checking CA injection for validating webhooks")
-			verifyCAInjection := func(g Gomega) {
+			Eventually(func(g Gomega) {
 				cmd := exec.Command("kubectl", "get",
 					"validatingwebhookconfigurations.admissionregistration.k8s.io",
 					"operator-validating-webhook-configuration",
 					"-o", "go-template={{ range .webhooks }}{{ .clientConfig.caBundle }}{{ end }}")
-				vwhOutput, err := utils.Run(cmd)
+				output, err := utils.Run(cmd)
 				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(len(vwhOutput)).To(BeNumerically(">", 10))
-			}
-			Eventually(verifyCAInjection).Should(Succeed())
+				g.Expect(len(output)).To(BeNumerically(">", 10))
+			}).Should(Succeed())
 		})
 
 		// +kubebuilder:scaffold:e2e-webhooks-checks
-
-		It("should create a Running pod for a Server", func() {
-			const testServerName = "e2e-server-test"
-
-			DeferCleanup(func() {
-				cmd := exec.Command("kubectl", "delete", "server", testServerName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-			})
-
-			By("applying a Server manifest")
-			tmpFile := writeServerManifest(testServerName, testNamespace,
-				`while true; do if wget -qO- http://localhost:8080/shutdown 2>/dev/null | grep -q 'true'; then wget -qO- --post-data='{"allowed":true}' http://localhost:8080/allow_delete 2>/dev/null; exit 0; fi; sleep 2; done`)
-			cmd := exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting for the pod to appear")
-			podName := testServerName + "-pod"
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "pod", podName,
-					"-n", testNamespace,
-					"-o", "jsonpath={.metadata.name}",
-				)
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(output).To(Equal(podName))
-			}, 2*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("waiting for the pod to reach Running phase")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "pod", podName,
-					"-n", testNamespace,
-					"-o", "jsonpath={.status.phase}",
-				)
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(output).To(Equal("Running"))
-			}, 3*time.Minute, 5*time.Second).Should(Succeed())
-		})
-
-		It("should inject the sidecar container into the Server pod", func() {
-			const testServerName = "e2e-server-sidecar-test"
-
-			DeferCleanup(func() {
-				cmd := exec.Command("kubectl", "delete", "server", testServerName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-			})
-
-			By("applying a Server manifest")
-			tmpFile := writeServerManifest(testServerName, testNamespace, "sleep 3600")
-			cmd := exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			podName := testServerName + "-pod"
-
-			By("waiting for the pod to appear")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "pod", podName,
-					"-n", testNamespace,
-					"-o", "jsonpath={.metadata.name}",
-				)
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(output).To(Equal(podName))
-			}, 2*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("verifying the sidecar container is present in the pod spec")
-			cmd = exec.Command("kubectl", "get", "pod", podName,
-				"-n", testNamespace,
-				"-o", "jsonpath={.spec.containers[*].name}",
-			)
-			output, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(output).To(ContainSubstring("fallernetes-sidecar"))
-		})
-
-		It("should remove the pod when a Server is deleted", func() {
-			const testServerName = "e2e-server-del-test"
-
-			DeferCleanup(func() {
-				cmd := exec.Command("kubectl", "delete", "server", testServerName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-			})
-
-			By("applying a Server manifest")
-			tmpFile := writeServerManifest(testServerName, testNamespace, "sleep 3600")
-			cmd := exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			podName := testServerName + "-pod"
-
-			By("waiting for the pod to exist")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "pod", podName,
-					"-n", testNamespace,
-					"-o", "jsonpath={.metadata.name}",
-				)
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(output).To(Equal(podName))
-			}, 2*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("deleting the Server")
-			cmd = exec.Command("kubectl", "delete", "server", testServerName, "-n", testNamespace, "--wait=false")
-			_, err = utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting for the pod to be gone")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "pod", podName, "-n", testNamespace)
-				_, err := utils.Run(cmd)
-				g.Expect(err).To(HaveOccurred())
-			}, 2*time.Minute, 5*time.Second).Should(Succeed())
-		})
-
-		It("should create a Fleet for a GameType", func() {
-			const testGTName = "e2e-gametype-test"
-
-			DeferCleanup(func() {
-				cmd := exec.Command("kubectl", "delete", "gametype", testGTName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-				Eventually(func(g Gomega) {
-					cmd := exec.Command("kubectl", "get", "fleets",
-						"-l", fmt.Sprintf("gametype=%s", testGTName),
-						"-n", testNamespace,
-						"-o", "jsonpath={.items[*].metadata.name}",
-					)
-					output, _ := utils.Run(cmd)
-					g.Expect(strings.TrimSpace(output)).To(BeEmpty())
-				}, 2*time.Minute, 5*time.Second).Should(Succeed())
-			})
-
-			By("applying a GameType manifest")
-			gtYAML := fmt.Sprintf(`apiVersion: gameserver.falloria.com/v1alpha1
-kind: GameType
-metadata:
-  name: %s
-  namespace: %s
-spec:
-  fleetSpec:
-    scaling:
-      replicas: 1
-      agePriority: oldest_first
-      prioritizeAllowed: false
-    spec:
-      sidecar:
-        port: 8080
-        image: %s
-      pod:
-        containers:
-        - name: game-server
-          image: busybox:latest
-          command: ["sh", "-c", "while true; do if wget -qO- http://localhost:8080/shutdown 2>/dev/null | grep -q 'true'; then wget -qO- --post-data='{\"allowed\":true}' http://localhost:8080/allow_delete 2>/dev/null; exit 0; fi; sleep 2; done"]
-`, testGTName, testNamespace, sidecarImage)
-			tmpFile := filepath.Join(os.TempDir(), "e2e-gametype.yaml")
-			Expect(os.WriteFile(tmpFile, []byte(gtYAML), 0644)).To(Succeed())
-			cmd := exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting for a Fleet labeled with the gametype name to appear")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "fleets",
-					"-l", fmt.Sprintf("gametype=%s", testGTName),
-					"-n", testNamespace,
-					"-o", "jsonpath={.items[*].metadata.name}",
-				)
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(utils.GetNonEmptyLines(output)).NotTo(BeEmpty())
-			}, 2*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("verifying the GameType status records the fleet name")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "gametype", testGTName,
-					"-n", testNamespace,
-					"-o", "jsonpath={.status.activeFleetName}",
-				)
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(output).NotTo(BeEmpty())
-			}, 2*time.Minute, 5*time.Second).Should(Succeed())
-		})
-
-		It("should create a new Fleet when GameType pod spec changes", func() {
-			const testGTName = "e2e-gametype-rolling-test"
-
-			DeferCleanup(func() {
-				cmd := exec.Command("kubectl", "delete", "gametype", testGTName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-				Eventually(func(g Gomega) {
-					cmd := exec.Command("kubectl", "get", "fleets",
-						"-l", fmt.Sprintf("gametype=%s", testGTName),
-						"-n", testNamespace,
-						"-o", "jsonpath={.items[*].metadata.name}",
-					)
-					output, _ := utils.Run(cmd)
-					g.Expect(strings.TrimSpace(output)).To(BeEmpty())
-				}, 2*time.Minute, 5*time.Second).Should(Succeed())
-			})
-
-			By("applying initial GameType")
-			gtYAML := fmt.Sprintf(`apiVersion: gameserver.falloria.com/v1alpha1
-kind: GameType
-metadata:
-  name: %s
-  namespace: %s
-spec:
-  fleetSpec:
-    scaling:
-      replicas: 1
-      agePriority: oldest_first
-      prioritizeAllowed: false
-    spec:
-      sidecar:
-        port: 8080
-        image: %s
-      pod:
-        containers:
-        - name: game-server
-          image: busybox:1.35
-          command: ["sh", "-c", "sleep 3600"]
-`, testGTName, testNamespace, sidecarImage)
-			tmpFile := filepath.Join(os.TempDir(), "e2e-gametype-rolling.yaml")
-			Expect(os.WriteFile(tmpFile, []byte(gtYAML), 0644)).To(Succeed())
-			cmd := exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting for the initial Fleet to appear")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "fleets",
-					"-l", fmt.Sprintf("gametype=%s", testGTName),
-					"-n", testNamespace,
-					"-o", "jsonpath={.items[*].metadata.name}",
-				)
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(utils.GetNonEmptyLines(output)).To(HaveLen(1))
-			}, 2*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("patching the GameType to change the container image")
-			cmd = exec.Command("kubectl", "patch", "gametype", testGTName,
-				"-n", testNamespace, "--type=merge",
-				"-p", `{"spec":{"fleetSpec":{"spec":{"pod":{"containers":[{"name":"game-server","image":"busybox:1.36","command":["sh","-c","sleep 3600"]}]}}}}}`,
-			)
-			_, err = utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting for the rolling update to produce a fleet with the new image")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "fleets",
-					"-l", fmt.Sprintf("gametype=%s", testGTName),
-					"-n", testNamespace,
-					"-o", "jsonpath={.items[*].spec.spec.pod.containers[*].image}",
-				)
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(output).To(ContainSubstring("busybox:1.36"))
-			}, 3*time.Minute, 5*time.Second).Should(Succeed())
-		})
-
-		It("should accept and reconcile a GameTypeAutoscaler without validation errors", func() {
-			const testASName = "e2e-autoscaler-test"
-			const testGTName = "e2e-autoscaler-gametype"
-
-			DeferCleanup(func() {
-				cmd := exec.Command("kubectl", "delete", "gametypeautoscaler", testASName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-				cmd = exec.Command("kubectl", "delete", "gametype", testGTName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-				Eventually(func(g Gomega) {
-					cmd := exec.Command("kubectl", "get", "fleets",
-						"-l", fmt.Sprintf("gametype=%s", testGTName),
-						"-n", testNamespace,
-						"-o", "jsonpath={.items[*].metadata.name}",
-					)
-					output, _ := utils.Run(cmd)
-					g.Expect(strings.TrimSpace(output)).To(BeEmpty())
-				}, 2*time.Minute, 5*time.Second).Should(Succeed())
-			})
-
-			By("creating a GameType for the autoscaler to target")
-			gtYAML := fmt.Sprintf(`apiVersion: gameserver.falloria.com/v1alpha1
-kind: GameType
-metadata:
-  name: %s
-  namespace: %s
-spec:
-  fleetSpec:
-    scaling:
-      replicas: 1
-      agePriority: oldest_first
-      prioritizeAllowed: false
-    spec:
-      sidecar:
-        port: 8080
-        image: %s
-      pod:
-        containers:
-        - name: game-server
-          image: busybox:latest
-          command: ["sh", "-c", "while true; do if wget -qO- http://localhost:8080/shutdown 2>/dev/null | grep -q 'true'; then wget -qO- --post-data='{\"allowed\":true}' http://localhost:8080/allow_delete 2>/dev/null; exit 0; fi; sleep 2; done"]
-`, testGTName, testNamespace, sidecarImage)
-			gtFile := filepath.Join(os.TempDir(), "e2e-autoscaler-gametype.yaml")
-			Expect(os.WriteFile(gtFile, []byte(gtYAML), 0644)).To(Succeed())
-			cmd := exec.Command("kubectl", "apply", "-f", gtFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("applying a GameTypeAutoscaler manifest")
-			// url points to an unreachable host; webhook call failures are expected and not tested here.
-			// Correct field names from the types: spec.policy (not autoscalePolicy), spec.sync.interval
-			// (not sync.time), and lowercase enum values: webhook / fixedinterval.
-			asYAML := fmt.Sprintf(`apiVersion: gameserver.falloria.com/v1alpha1
-kind: GameTypeAutoscaler
-metadata:
-  name: %s
-  namespace: %s
-spec:
-  gameTypeName: %s
-  policy:
-    type: webhook
-    webhook:
-      url: http://localhost:9999
-      path: scale
-  sync:
-    type: fixedinterval
-    interval: 30s
-`, testASName, testNamespace, testGTName)
-			tmpFile := filepath.Join(os.TempDir(), "e2e-autoscaler.yaml")
-			Expect(os.WriteFile(tmpFile, []byte(asYAML), 0644)).To(Succeed())
-			cmd = exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err = utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred(), "GameTypeAutoscaler should be accepted by the webhook")
-
-			By("verifying the resource exists in the cluster")
-			Eventually(func(g Gomega) {
-				cmd := exec.Command("kubectl", "get", "gametypeautoscaler", testASName,
-					"-n", testNamespace, "-o", "jsonpath={.metadata.name}")
-				output, err := utils.Run(cmd)
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(output).To(Equal(testASName))
-			}, time.Minute, 5*time.Second).Should(Succeed())
-
-			By("checking that no validation error events were emitted for the autoscaler")
-			// GameautoscalerWebhook warnings are expected (no server in e2e); only check that the
-			// CR fields were accepted as valid by the controller.
-			validationReasons := []string{
-				"GameAutoscalerInvalidTarget",
-				"GameautoscalerInvalidAutoscalePolicy",
-				"GameautoscalerInvalidSyncType",
-			}
-			Consistently(func(g Gomega) {
-				for _, reason := range validationReasons {
-					cmd := exec.Command("kubectl", "get", "events",
-						"-n", testNamespace,
-						"--field-selector", fmt.Sprintf("involvedObject.name=%s,reason=%s", testASName, reason),
-						"-o", "jsonpath={.items[*].reason}",
-					)
-					output, _ := utils.Run(cmd)
-					g.Expect(output).To(BeEmpty())
-				}
-			}, 10*time.Second, 2*time.Second).Should(Succeed())
-		})
-
-		It("should create Running pods for a Fleet", func() {
-			const testFleetName = "e2e-fleet-test"
-
-			DeferCleanup(func() {
-				cmd := exec.Command("kubectl", "delete", "fleet", testFleetName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-			})
-
-			By("applying a sample Fleet manifest")
-			tmpFile := writeFleetManifest(testFleetName, testNamespace, 2)
-			cmd := exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred(), "Failed to apply Fleet manifest")
-
-			By("waiting for 2 pods to appear with the fleet label")
-			Eventually(waitForFleetPods(testFleetName, testNamespace, 2), 2*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("verifying all pods reach Running phase")
-			Eventually(waitForFleetPodsRunning(testFleetName, testNamespace), 3*time.Minute, 5*time.Second).Should(Succeed())
-		})
-
-		It("should remove all Servers and Pods when a Fleet is deleted", func() {
-			const testFleetName = "e2e-fleet-del-test"
-
-			By("creating a Fleet with 2 replicas")
-			tmpFile := writeFleetManifest(testFleetName, testNamespace, 2)
-			cmd := exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred(), "Failed to apply Fleet manifest")
-
-			By("waiting for 2 pods to be Running")
-			Eventually(waitForFleetPods(testFleetName, testNamespace, 2), 2*time.Minute, 5*time.Second).Should(Succeed())
-			Eventually(waitForFleetPodsRunning(testFleetName, testNamespace), 3*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("deleting the Fleet")
-			cmd = exec.Command("kubectl", "delete", "fleet", testFleetName, "-n", testNamespace, "--wait=false")
-			_, err = utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting for all Pods with the fleet label to disappear")
-			Eventually(waitForFleetPods(testFleetName, testNamespace, 0), 3*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("waiting for all Server CRs with the fleet label to disappear")
-			Eventually(waitForFleetServers(testFleetName, testNamespace, 0), 3*time.Minute, 5*time.Second).Should(Succeed())
-		})
-
-		It("should remove excess pods when a Fleet is scaled down", func() {
-			const testFleetName = "e2e-fleet-scaledown-test"
-
-			DeferCleanup(func() {
-				cmd := exec.Command("kubectl", "delete", "fleet", testFleetName, "-n", testNamespace,
-					"--wait=false", "--ignore-not-found=true")
-				_, _ = utils.Run(cmd)
-			})
-
-			By("applying a Fleet with 3 replicas")
-			tmpFile := writeFleetManifest(testFleetName, testNamespace, 3)
-			cmd := exec.Command("kubectl", "apply", "-f", tmpFile)
-			_, err := utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting for 3 pods to appear")
-			Eventually(waitForFleetPods(testFleetName, testNamespace, 3), 2*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("patching fleet replicas down to 1")
-			cmd = exec.Command("kubectl", "patch", "fleet", testFleetName,
-				"-n", testNamespace,
-				"--type=merge",
-				"-p", `{"spec":{"scaling":{"replicas":1}}}`,
-			)
-			_, err = utils.Run(cmd)
-			Expect(err).NotTo(HaveOccurred())
-
-			By("waiting until only 1 pod remains")
-			Eventually(waitForFleetPods(testFleetName, testNamespace, 1), 2*time.Minute, 5*time.Second).Should(Succeed())
-
-			By("waiting until only 1 Server CR remains")
-			Eventually(waitForFleetServers(testFleetName, testNamespace, 1), 2*time.Minute, 5*time.Second).Should(Succeed())
-		})
-
 	})
+
+	Context("Server", serverTests)
+	Context("GameType", gameTypeTests)
+	Context("Fleet", fleetTests)
 })
-
-// serviceAccountToken returns a token for the specified service account in the given namespace.
-// It uses the Kubernetes TokenRequest API to generate a token by directly sending a request
-// and parsing the resulting token from the API response.
-func serviceAccountToken() (string, error) {
-	const tokenRequestRawString = `{
-		"apiVersion": "authentication.k8s.io/v1",
-		"kind": "TokenRequest"
-	}`
-
-	By("creating temporary file to store the token request")
-	secretName := fmt.Sprintf("%s-token-request", serviceAccountName)
-	tokenRequestFile := filepath.Join("/tmp", secretName)
-	err := os.WriteFile(tokenRequestFile, []byte(tokenRequestRawString), os.FileMode(0o644))
-	if err != nil {
-		return "", err
-	}
-
-	var out string
-	verifyTokenCreation := func(g Gomega) {
-		By("executing kubectl command to create the token")
-		cmd := exec.Command("kubectl", "create", "--raw", fmt.Sprintf(
-			"/api/v1/namespaces/%s/serviceaccounts/%s/token",
-			namespace,
-			serviceAccountName,
-		), "-f", tokenRequestFile)
-
-		output, err := cmd.CombinedOutput()
-		g.Expect(err).NotTo(HaveOccurred())
-
-		By("parsing the JSON output to extract the token")
-		var token tokenRequest
-		err = json.Unmarshal(output, &token)
-		g.Expect(err).NotTo(HaveOccurred())
-
-		out = token.Status.Token
-	}
-	Eventually(verifyTokenCreation).Should(Succeed())
-
-	return out, err
-}
-
-// getMetricsOutput retrieves and returns the logs from the curl pod used to access the metrics endpoint.
-func getMetricsOutput() (string, error) {
-	By("getting the curl-metrics logs")
-	cmd := exec.Command("kubectl", "logs", "curl-metrics", "-n", namespace)
-	return utils.Run(cmd)
-}
-
-// tokenRequest is a simplified representation of the Kubernetes TokenRequest API response,
-// containing only the token field that we need to extract.
-type tokenRequest struct {
-	Status struct {
-		Token string `json:"token"`
-	} `json:"status"`
-}
-
-// waitForFleetPods returns a Gomega assertion function that passes when exactly count pods
-// with the given fleet label exist. Use count=0 to assert all pods are gone.
-func waitForFleetPods(fleetName, ns string, count int) func(Gomega) {
-	return func(g Gomega) {
-		cmd := exec.Command("kubectl", "get", "pods",
-			"-l", fmt.Sprintf("fleet=%s", fleetName),
-			"-n", ns,
-			"-o", "jsonpath={.items[*].metadata.name}")
-		output, err := utils.Run(cmd)
-		g.Expect(err).NotTo(HaveOccurred())
-		g.Expect(strings.Fields(output)).To(HaveLen(count))
-	}
-}
-
-// waitForFleetServers returns a Gomega assertion function that passes when exactly count Server CRs
-// with the given fleet label exist. Use count=0 to assert all servers are gone.
-func waitForFleetServers(fleetName, ns string, count int) func(Gomega) {
-	return func(g Gomega) {
-		cmd := exec.Command("kubectl", "get", "servers",
-			"-l", fmt.Sprintf("fleet=%s", fleetName),
-			"-n", ns,
-			"-o", "jsonpath={.items[*].metadata.name}")
-		output, err := utils.Run(cmd)
-		g.Expect(err).NotTo(HaveOccurred())
-		g.Expect(strings.Fields(output)).To(HaveLen(count))
-	}
-}
-
-// waitForFleetPodsRunning returns a Gomega assertion function that passes when every pod
-// with the given fleet label is in the Running phase.
-func waitForFleetPodsRunning(fleetName, ns string) func(Gomega) {
-	return func(g Gomega) {
-		cmd := exec.Command("kubectl", "get", "pods",
-			"-l", fmt.Sprintf("fleet=%s", fleetName),
-			"-n", ns,
-			"-o", "jsonpath={.items[*].status.phase}")
-		output, err := utils.Run(cmd)
-		g.Expect(err).NotTo(HaveOccurred())
-		for _, phase := range strings.Fields(output) {
-			g.Expect(phase).To(Equal("Running"))
-		}
-	}
-}
-
-// writeServerManifest writes a Server manifest to a temp file and returns its path.
-func writeServerManifest(name, ns, command string) string {
-	escapedCmd := strings.ReplaceAll(command, `"`, `\"`)
-	yaml := fmt.Sprintf(`apiVersion: gameserver.falloria.com/v1alpha1
-kind: Server
-metadata:
-  name: %s
-  namespace: %s
-spec:
-  sidecar:
-    port: 8080
-    image: %s
-  pod:
-    containers:
-    - name: game-server
-      image: busybox:latest
-      command: ["sh", "-c", "%s"]
-`, name, ns, sidecarImage, escapedCmd)
-	path := filepath.Join(os.TempDir(), name+".yaml")
-	Expect(os.WriteFile(path, []byte(yaml), 0644)).To(Succeed())
-	return path
-}
-
-// writeFleetManifest writes a Fleet manifest to a temp file and returns its path.
-// No force-delete timeout is set: deletion must be approved via the sidecar protocol so that
-// tests catch regressions in the sidecar communication path.
-func writeFleetManifest(name, ns string, replicas int) string {
-	filename := name + ".yaml"
-	yaml := fmt.Sprintf(`apiVersion: gameserver.falloria.com/v1alpha1
-kind: Fleet
-metadata:
-  name: %s
-  namespace: %s
-spec:
-  scaling:
-    replicas: %d
-    agePriority: oldest_first
-    prioritizeAllowed: false
-  spec:
-    sidecar:
-      port: 8080
-      image: %s
-    pod:
-      containers:
-      - name: game-server
-        image: busybox:latest
-        command: ["sh", "-c", "while true; do if wget -qO- http://localhost:8080/shutdown 2>/dev/null | grep -q 'true'; then wget -qO- --post-data='{\"allowed\":true}' http://localhost:8080/allow_delete 2>/dev/null; exit 0; fi; sleep 2; done"]
-`, name, ns, replicas, sidecarImage)
-	path := filepath.Join(os.TempDir(), filename)
-	Expect(os.WriteFile(path, []byte(yaml), 0644)).To(Succeed())
-	return path
-}
